refactor: Use `vulture` to remove dead code
