+++ neps/sampling/__init__.py
@@ -1,4 +1,4 @@
-from neps.sampling.priors import CenteredPrior, Prior, UniformPrior, WeightedPrior
+from neps.sampling.priors import CenteredPrior, Prior, UniformPrior
 from neps.sampling.samplers import Sampler, Sobol
 
-__all__ = ["Sobol", "Sampler", "Prior", "UniformPrior", "CenteredPrior", "WeightedPrior"]
+__all__ = ["Sobol", "Sampler", "Prior", "UniformPrior", "CenteredPrior"]
