+++ neps/search_spaces/architecture/cfg_variants/constrained_cfg.py
@@ -1,9 +1,7 @@
 from __future__ import annotations
 
-import itertools
 import math
 from collections import deque
-from copy import deepcopy
 from functools import partial
 from queue import LifoQueue
 
@@ -27,9 +25,6 @@ class Constraint:
     def update_context(self, new_part: str) -> None:
         raise NotImplementedError
 
-    def get_all_potential_productions(self, production) -> list:
-        raise NotImplementedError
-
     def mutate_not_allowed_productions(
         self, nonterminal: str, before: str, after: str, possible_productions: list
     ) -> list:
@@ -37,22 +32,18 @@ class Constraint:
 
 
 class ConstrainedGrammar(Grammar):
+    def set_constraints(self, constraints: dict, none_operation: str | None = None):
+        self.constraints = constraints
+        self.none_operation = none_operation
+        self.constraint_is_class = isinstance(self.constraints, Constraint)
+
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.constraints = None
         self.none_operation = None
         self.constraint_is_class: bool = False
 
-        self._prior: dict = None
-
-    def set_constraints(self, constraints: dict, none_operation: str | None = None):
-        self.constraints = constraints
-        self.none_operation = none_operation
-        self.constraint_is_class = isinstance(self.constraints, Constraint)
-
-    @staticmethod
-    def is_constrained():
-        return True
+        self._prior: dict | None = None
 
     @property
     def prior(self):
@@ -393,7 +384,7 @@ class ConstrainedGrammar(Grammar):
                     q_context.pop()
                     s = s[:-1]
                 if len(q_context) == 0:
-                    tmp_str = self._remove_empty_spaces(tmp_str)
+                    tmp_str = tmp_str.strip()
                     current_derivation.append(tmp_str)
                     if len(productions) == 1 and len(current_derivation) == len(
                         self.constraints(productions[0].rhs()[0])
@@ -435,7 +426,7 @@ class ConstrainedGrammar(Grammar):
                             q_context.pop()
                         s = s[:-1]
                     if len(q_context) == 0:
-                        tmp_str = self._remove_empty_spaces(tmp_str)
+                        tmp_str = tmp_str.strip()
                         current_derivation.append(tmp_str)
                         if len(productions) == 1 and len(current_derivation) == len(
                             self.constraints(productions[0].rhs()[0])
@@ -502,7 +493,7 @@ class ConstrainedGrammar(Grammar):
             ):
                 break
             _patience -= 1
-        return self._remove_empty_spaces(child)
+        return child.strip()
 
     def crossover(
         self,
@@ -605,8 +596,8 @@ class ConstrainedGrammar(Grammar):
                     # return the two new tree
                     child1 = pre + donor_sub + post
                     child2 = donor_pre + sub + donor_post
-                    child1 = self._remove_empty_spaces(child1)
-                    child2 = self._remove_empty_spaces(child2)
+                    child1 = child1.strip()
+                    child2 = child2.strip()
 
                     if return_crossover_subtrees:
                         return (
@@ -618,132 +609,3 @@ class ConstrainedGrammar(Grammar):
                     return child1, child2
 
         raise Exception("Cannot do crossover")
-
-    @property
-    def compute_space_size(self) -> int:
-        """Computes the size of the space described by the grammar.
-
-        Args:
-            primitive_nonterminal (str, optional): The primitive nonterminal of the grammar. Defaults to "OPS".
-
-        Returns:
-            int: size of space described by grammar.
-        """
-
-        def recursive_worker(
-            nonterminal: Nonterminal, memory_bank: dict | None = None
-        ) -> int:
-            def _get_all_variants(production):
-                variants = [production]
-                nonterminals = [
-                    i
-                    for i, sym in enumerate(production.rhs())
-                    if isinstance(sym, Nonterminal)
-                ]
-                max_zero_op = len(nonterminals)
-                for n_zero_op in range(1, max_zero_op):
-                    for zero_combination in itertools.combinations(
-                        nonterminals, n_zero_op
-                    ):
-                        current_derivation = self.constraints(production.rhs()[0])
-                        counter = 0
-                        valid_production = True
-                        for i, sym in enumerate(production.rhs()):
-                            if not isinstance(sym, str):
-                                context_information = self.constraints(
-                                    production.rhs()[0],
-                                    current_derivation,
-                                )
-                                if isinstance(context_information, list):
-                                    not_allowed_productions = (
-                                        self._get_not_allowed_productions(
-                                            self.productions(lhs=sym),
-                                            context_information[counter],
-                                        )
-                                    )
-                                elif isinstance(context_information, bool):
-                                    not_allowed_productions = (
-                                        self._get_not_allowed_productions(
-                                            self.productions(lhs=sym),
-                                            context_information,
-                                        )
-                                    )
-                                else:
-                                    raise NotImplementedError
-                                if (
-                                    i in zero_combination
-                                    and len(not_allowed_productions) > 0
-                                ):
-                                    valid_production = False
-                                    break
-                                if i >= max(zero_combination):
-                                    break
-                                if i in zero_combination:
-                                    current_derivation[counter] = self.none_operation
-                                counter += 1
-                        if valid_production:
-                            new_production = deepcopy(production)
-                            rhs = list(new_production.rhs())
-                            new_production._rhs = tuple(
-                                self.none_operation if i in zero_combination else r
-                                for i, r in enumerate(rhs)
-                            )
-                            variants.append(new_production)
-                return variants
-
-            if memory_bank is None:
-                memory_bank = {}
-
-            _potential_productions = self.productions(lhs=nonterminal)
-            potential_productions = []
-            for potential_production in _potential_productions:
-                nonterminals = list(
-                    {
-                        sym
-                        for sym in potential_production.rhs()
-                        if isinstance(sym, Nonterminal)
-                    }
-                )
-                if self.constraint_is_class:
-                    potential_productions += (
-                        self.constraints.get_all_potential_productions(
-                            potential_production
-                        )
-                    )
-                elif any(
-                    production.rhs()[0] == self.none_operation
-                    for nonterminal in nonterminals
-                    for production in self.productions(nonterminal)
-                ):
-                    potential_productions += _get_all_variants(potential_production)
-                elif not (
-                    len(potential_production.rhs()) == 1
-                    and potential_production.rhs()[0] == self.none_operation
-                ):
-                    potential_productions.append(potential_production)
-            _possibilites = 0
-            for potential_production in potential_productions:
-                nonterminals = [
-                    rhs_sym
-                    for rhs_sym in potential_production.rhs()
-                    if isinstance(rhs_sym, Nonterminal)
-                ]
-                possibilities_per_edge = [
-                    memory_bank[str(e_nonterminal)]
-                    if str(e_nonterminal) in memory_bank
-                    else recursive_worker(e_nonterminal, memory_bank)
-                    for e_nonterminal in nonterminals
-                ]
-                memory_bank.update(
-                    {
-                        str(e_nonterminal): possibilities_per_edge[i]
-                        for i, e_nonterminal in enumerate(nonterminals)
-                    }
-                )
-                product = 1
-                for p in possibilities_per_edge:
-                    product *= p
-                _possibilites += product
-            return _possibilites
-
-        return recursive_worker(self.start())
