+++ neps/search_spaces/architecture/mutations.py
@@ -68,63 +68,3 @@ def bananas_mutate(  # noqa: D103
         idx += 1
 
     return child_string_tree, child_string_tree == parent_string_tree
-
-
-def repetitive_search_space_mutation(  # noqa: D103
-    base_parent: str,
-    motif_parents: list[str],
-    base_grammar: Grammar,
-    motif_grammars: list[Grammar],
-    terminal_to_sublanguage_map: dict,
-    number_of_repetitive_motifs_per_grammar: list,
-    inner_mutation_strategy: Callable,
-    mutation_rate: float = 1.0,
-    mutation_prob: float | None = None,
-    fixed_macro_parent: bool = False,  # noqa: FBT001, FBT002
-) -> tuple[list[str], list[bool]]:
-    def _motifs_in_base_tree(base_parent, terminal_to_sublanguage_map):
-        return [
-            i
-            for i, k in enumerate(terminal_to_sublanguage_map.keys())
-            if k in base_parent
-        ]
-
-    indices = _motifs_in_base_tree(base_parent, terminal_to_sublanguage_map)
-    if fixed_macro_parent:
-        mutation_prob = (
-            mutation_rate / len(indices) if mutation_prob is None else mutation_prob
-        )
-    else:
-        mutation_prob = (
-            mutation_rate / (len(indices) + 1) if mutation_prob is None else mutation_prob
-        )
-
-    child_string_trees = []
-    if not fixed_macro_parent and random.random() < mutation_prob:  # noqa: S311
-        child_string_trees.append(inner_mutation_strategy(base_parent, base_grammar))
-        indices = _motifs_in_base_tree(base_parent, terminal_to_sublanguage_map)
-        mutation_prob = (
-            mutation_rate / (len(indices) + 1) if mutation_prob is None else mutation_prob
-        )
-    else:
-        child_string_trees.append((base_parent, True))
-
-    parent_string_idx = 0
-    _number_of_repetitive_motifs_per_grammar = (
-        number_of_repetitive_motifs_per_grammar[1:]
-        if not fixed_macro_parent
-        else number_of_repetitive_motifs_per_grammar
-    )
-    for grammar, number_of_motifs in zip(
-        motif_grammars, _number_of_repetitive_motifs_per_grammar
-    ):
-        for _ in range(number_of_motifs):
-            if parent_string_idx in indices and random.random() < mutation_prob:  # noqa: S311
-                child_string_trees.append(
-                    inner_mutation_strategy(motif_parents[parent_string_idx], grammar)
-                )
-            else:
-                child_string_trees.append((motif_parents[parent_string_idx], True))
-            parent_string_idx += 1
-
-    return [c[0] for c in child_string_trees], [c[1] for c in child_string_trees]
