+++ tests/test_samplers.py
@@ -1,6 +1,6 @@
 from pytest_cases import parametrize
 from neps.sampling.samplers import Sampler, Sobol, WeightedSampler, BorderSampler
-from neps.sampling.priors import Prior, UniformPrior, WeightedPrior
+from neps.sampling.priors import Prior, UniformPrior
 
 import torch
 
@@ -24,10 +24,6 @@ def _make_centered_prior(ndim: int) -> Prior:
         _make_centered_prior(ndim=3),
         WeightedSampler(
             [UniformPrior(ndim=3), _make_centered_prior(3), Sobol(ndim=3)],
-            weights=torch.tensor([0.5, 0.25, 0.25]),
-        ),
-        WeightedPrior(
-            [UniformPrior(ndim=3), _make_centered_prior(3), UniformPrior(ndim=3)],
             weights=torch.tensor([0.5, 0.25, 0.25]),
         ),
     ],
@@ -63,10 +59,6 @@ def test_sampler_samples_into_domain(sampler: Sampler) -> None:
         UniformPrior(ndim=3),
         # Convenence method for making a distribution around center points
         _make_centered_prior(ndim=3),
-        WeightedPrior(
-            [UniformPrior(ndim=3), _make_centered_prior(3), UniformPrior(ndim=3)],
-            weights=torch.tensor([0.5, 0.25, 0.25]),
-        ),
     ],
 )
 def test_priors_give_positive_pdfs(prior: Prior) -> None:
